unchecked on all code, saves huge amount of gas
went from 2359896 to 1363870 on the function call w/ provided test data, nearly a million gas saved, or 42% less
also saved almost 200k gas on deploy (from 2037981 down to 1841982)
i saw the "note: solidity reverts on underflow", and thought they were using it as a feature, but not really, since it's already checking if len is 0 and it's a uint, so the check is already there; no checked code blocks needed
i wonder how much gas solidity is wasting overall post-0.8

also, really? 100 lines of numbers for an array constant? cmon